+++ src/java/android/telephony/SmsManager.java
@@ -65,10 +65,10 @@ public final class SmsManager {
     /** Singleton object constructed during class initialization. */
     private static final SmsManager sInstance = new SmsManager(DEFAULT_SUB_ID);
     private static final Object sLockObject = new Object();
-    private static final Map<Long, SmsManager> sSubInstances = new ArrayMap<Long, SmsManager>();
+    private static final Map<Integer, SmsManager> sSubInstances = new ArrayMap<Integer, SmsManager>();
 
     /** A concrete subId, or the pseudo DEFAULT_SUB_ID */
-    private long mSubId;
+    private int mSubId;
 
     /*
      * Key for the various carrier-dependent configuration values.
@@ -710,7 +710,7 @@ public final class SmsManager {
      *
      * {@hide}
      */
-    public static SmsManager getSmsManagerForSubscriber(long subId) {
+    public static SmsManager getSmsManagerForSubscriber(int subId) {
         // TODO(shri): Add javadoc link once SubscriptionManager is made public api
         synchronized(sLockObject) {
             SmsManager smsManager = sSubInstances.get(subId);
@@ -722,7 +722,7 @@ public final class SmsManager {
         }
     }
 
-    private SmsManager(long subId) {
+    private SmsManager(int subId) {
         mSubId = subId;
     }
 
@@ -755,8 +755,7 @@ public final class SmsManager {
      *
      * {@hide}
      */
-    public long getSubId() {
-        // TODO(shri): Add javadoc link once SubscriptionManager is made public api
+    public int getSubId() {
         if (mSubId == DEFAULT_SUB_ID) {
             return getDefaultSmsSubId();
         }
@@ -1153,7 +1152,7 @@ public final class SmsManager {
      * @return the default SubId
      * @hide
      */
-    public static long getDefaultSmsSubId() {
+    public static int getDefaultSmsSubId() {
         return SubscriptionManager.getDefaultSmsSubId();
     }
